\name{statistica-package}
\alias{statistica-package}
\alias{statistica}
\docType{package}
\title{
\packageTitle{statistica}
}
\description{
\packageDescription{statistica}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{statistica}
\packageIndices{statistica}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{statistica}

Maintainer: \packageMaintainer{statistica}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
